fix: rename webview-auto-height; fix RN version
