+++ packages/gpt/package.json
@@ -49,7 +49,7 @@
     "react": "16.0.0-alpha.6",
     "react-broadcast": "^0.1.2",
     "react-dom": "15.5.4",
-    "react-native": "^0.46.0"
+    "react-native": "^0.44.1"
   },
   "peerDependencies": {
     "@storybook/react-native": "3.1.6"
